feat: add apipass subcommand
Used to check if apicheck passes

Log:
