+++ main.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"io"
 	"log"
 	"net/http"
 	"os"
@@ -466,70 +465,6 @@ func (cl *Client) SetupCloseHandler() {
 		cl.GetApiJobCancel()
 		os.Exit(0)
 	}()
-}
-
-func subcommands(args []string) bool {
-	if len(os.Args) < 2 {
-		return false
-	}
-	cl := NewClient()
-	switch os.Args[1] {
-	case "wait": // wait job end
-		var runid int
-		var server, token string
-		waitFlag := flag.NewFlagSet("wait", flag.ExitOnError)
-		waitFlag.IntVar(&runid, "runid", 0, "job runid")
-		waitFlag.StringVar(&server, "server", "", "server")
-		waitFlag.StringVar(&token, "token", "", "token (env BRIDGE_TOKEN)")
-		waitFlag.Parse(os.Args[2:])
-		if len(token) == 0 {
-			token = os.Getenv("BRIDGE_TOKEN")
-		}
-		if runid == 0 {
-			waitFlag.Usage()
-			os.Exit(1)
-		}
-		cl.host = server
-		cl.token = token
-		cl.id = runid
-		err := cl.Wait()
-		if err != nil {
-			log.Fatal(err)
-		}
-		return true
-	case "cat": // print job artifact
-		var runid int
-		var server, token string
-		var filename string
-		catFlag := flag.NewFlagSet("cat", flag.ExitOnError)
-		catFlag.IntVar(&runid, "runid", 0, "job runid")
-		catFlag.StringVar(&filename, "file", "", "file path")
-		catFlag.StringVar(&server, "server", "", "server")
-		catFlag.StringVar(&token, "token", "", "token (env BRIDGE_TOKEN)")
-		catFlag.Parse(os.Args[2:])
-		if len(token) == 0 {
-			token = os.Getenv("BRIDGE_TOKEN")
-		}
-		if runid == 0 || len(filename) == 0 {
-			catFlag.Usage()
-			os.Exit(1)
-		}
-		cl.host = server
-		cl.token = token
-		cl.id = runid
-		for _, f := range cl.GetApiJobArtifacts().Files {
-			if f.Name == filename {
-				resp, err := cl.R().SetDoNotParseResponse(true).Get(f.URL)
-				if err != nil {
-					log.Fatal(err)
-				}
-				defer resp.RawBody().Close()
-				io.Copy(os.Stdout, resp.RawBody())
-			}
-		}
-		return true
-	}
-	return false
 }
 
 func main() {
